+++ src/search/mod.rs
@@ -1,0 +1,11 @@
+
+mod fuzzy_pattern;
+mod name_match;
+mod pos;
+
+pub use {
+    fuzzy_pattern::FuzzyPattern,
+    name_match::NameMatch,
+    pos::*,
+};
+