+++ src/search/pos.rs
@@ -1,0 +1,5 @@
+
+/// a vector of indexes of the matching characters (not bytes)
+pub type Pos = Vec<usize>;
+
+