+++ src/tui/entry_ref.rs
@@ -1,0 +1,10 @@
+
+
+/// a type used for entry display, valable when there
+/// is or there isn't a search
+pub struct EntryRef<'d, 's> {
+
+    /// idx of the entry in the drawer
+    pub idx: usize,
+
+}